BUG/MINOR: sink: add tempo between 2 connection attempts for sft servers (2)
This is a follow up patch for 9e02e4d5 ("BUG/MINOR: sink: add tempo
between 2 connection attempts for sft servers").

When the patch was backported to 3.0, due to context mismatch it had to
be manually adjusted but a mistake was made. Indeed, the fix is a no-op
given that the last_conn is always set to TICK_ETERNITY (which translates
to 0) that corresponds to the starting value (sft is calloced) where the
applet is allowed to be created instantly. Instead, we must set last_conn
to now_ms as it was done in the upstream patch.

It should be backported everywhere 9e02e4d5 was.

This commit has no upstream id because it is specific to the 3.0

(cherry picked from commit 87f3c100f190c97a21d362bf04d8339b79c9ca8b)
Signed-off-by: Aurelien DARRAGON <adarragon@haproxy.com>